Extract SubtitleBar into its own file
